Use the correct format specifier for size_t in fz_strdup.
Prior to this patch, if the strdup failed, it would throw an error
printing the number of bytes that could not be allocated. However,
this was formatted as an int, while the actual argument passed is
the return value of strdup itself, which is usually a size_t.

This patch formats instead as an unsigned long, and adds a cast to
ensure the types match even on platforms where this isn't already
the case.