Cast pointers to unsigned long to avoid truncation.
In these two places, a pointer's alignment was checked by casting it
to an unsigned in and then masking off all but the last low two bits.

This method generates a warning on LP64 systems like Linux x86_64 since
the pointer must be truncated. This is fine as far as the algorithm is
concerned, but gcc generates a warning which we would like to remove.

Casting to (unsigned long) resolves the warning, since this is the
same width as a pointer on Linux x86_64. On 32 bit machines, it is
generally also the same width as a pointer, and as the previous
(unsigned) cast. However, it won't help on Windows x86_64, where
long is still 32 bits.