More portable fix for pointer alignments.
The previous patch cast the pointer to (unsigned long) which fixed
a truncation warning on 64 bit Linux, but didn't help on Win64.

Casting to (size_t) is a little more portable, but mupdf doesn't
use that type. Casting to (char*) and substractiong (char*)0
is the best solution for this. It looks a little messy; if more
such checks are added, I suggest moving to a macro:

#define PTR_ALIGN4(ptr) (!((char *)(ptr) - (char *)0) & 3))

However, the pointer in rendersolid() is already an (unsigned) char
pointer. The check in decodetile() is actually checking an unsigned int
pointer for 4-octet alignment, so I don't see how that can ever fail.