Makefile tweaks for cross compiling, plus example cross compile target.
Add 2 new makefile options to Mupdf. If CROSSCOMPILE is defined, then we
avoid performing tasks during the build that require a binary to be built
and then executed as part of the build. Currently this is just the cmap
and font dumping steps.

If NOX11 is defined, then we avoid building the X11 app.

Finally, in Makerules, we have a new section to show how to encapsulate
the changes for a given cross compile target. If OS is defined to be
"beagle-cross" then we build using given compilers/options.