Non-isolated group support, and fix Bug 692336.
Firstly, this takes on some of Zenikos patch to correct the clip
stack handling that was broken by the fix to bug 692287 (in commit
2c3bbbf). This bug should now be solved.

We add a new 'shape' field to the draw device structure (and clip
stack). When we are inside non-isolated groups, this is set to be
a pixmap where we accumulate the 'shape' of the objects drawn.

When we come to blend back, if we are blending a non-isolated group
back, we have to use a different blending function that takes account
of the shape.

Various internal groups (the page group, and groups used to force
blending) are set to be isolated to avoid carrying shape planes
around when this is not required.

All our rendering code now has to know how to maintain the shape
plane as well as doing the basic rendering.