Add support for colored Type3 glyphs (ones that use d0).
Detect the d0 or d1 operators by writing a bit to the new device
flags word. This can then be checked by the Type3 code to create
the appropriate backing pixmap.

In order to know what the appropriate backing pixmap is, we pass
an additional colorspace into the glyph rendering code.