ARM opts in draw_scale.c; updated android build opts.
Kammerer reports 90%+ of CPU time is spent in the image scaling code
for his documents on Android.

In this commit we provide ARM optimised cores for the common scaling
routines (1/2/4 components). Tests indicate this doubles the speed of
rendering for a bitmap heavy PDF file on an HTC desire.

This code is included if ARCH_ARM is defined. If ARCH_THUMB is defined
then extra instructions are added to ensure correct interworking.

We also update the Android jni makefiles to set these defines.

We update the ReadMe.txt with more explicit instructions and update with
more modern ndk/sdk versions.

We update build.xml in line with new sdk releases.