Force MuPageView dealloc to always occur in main thread.
When a page view was removed (because it scrolled off too far) while
the background thread had its page queued, when the block completed
and released the page view in the cancel case (thus not passing a
block back to the main queue), the dealloc would trigger in the
worker thread. This is not safe to do with UIKit, causing crashes!