Fix bug 692627; pdfclean -ggg of encrypted files fails.
Encryption bakes the object numbers into the strings/streams. This
means that our renumbering technique falls down; when we move an
object to its new position, then come to read it, we find that it
decrypts badly.

The fix here is (as suggested by Zeniko) to avoid renumbering
when encryption is in use.