First cut at support for OCGs in mupdf (bug 692314)
When opening a file, create a pdf_ocg_descriptor that lists the OCGs
in a file. Add a new function to allow us to set the configuration
in use (currently just the default one).

This sets the states of the OCGs as appropriate. When decoding the
file respect the states of the OCGs.

This results in Invite.pdf rendering correctly.

There is more to be done in this area (with automatic setting of
OCGs by language/zoom level etc), but this is a good start.