Fix lack of output with "719 - EOF incorrectly detected.pdf"
When converting to exception handling I'd messed up an error handling
case; when failing to pdf_lex in pdf_repair_xref I had allowed the
error to just carry on being thrown rather than catching it and
cleaning up. This was resulting in not getting any output for the
above file, rather than outputting as much as we could.

Simple fix.