Add scavenging functionality.
When fz_malloc (etc) are about to fail, we try to scavenge memory
from the store and then retry. We repeatedly try to bin objects
from the store until the malloc succeeds, or until we have nothing
else to bin.

This means we no longer need the 'aging' of the store, so this is
removed.