New Memento features; nesting of listed blocks, labelling of blocks
When listing blocks from Memento, look for internal pointers to
other blocks, and use this to generate a nested list - much
easier to read when looking to see which stray reference is
causing blocks to leak.

Add new Memento_label feature to allow blocks to be labelled. Makes
block displaying much easier to read.