Rework draw device stack handling.
As we create clips/transparency groups etc, we maintain a stack of
draw device states.

Here we change this from having a 'current' state, and storing changes
on the stack to always keeping the complete current state on the head
of the stack. This should make error cleanup easier, as well as being
conceptually easier to follow (for me at least).