Add 'marking' functions for dictionaries.
In various places in the code, we add markers (".seen") to
dictionaries as we traverse them to ensure that we don't
go into infinite loops.

Adding a dictionary entry is bad as it's a) an expensive
operation, b) a potentially destructive one, and c) produces another
possible point of failure (as mallocs can fail).

Instead, add a flag to each dict to allow them to be marked/unmarked
and use that instead.

Thanks to Zeniko for pointing out various places that could usefully
be protected against infinite recursion.