Add fz_blend_pixel_function.
Inspired by Sumatras gdi devices need for such a function expose
the blending functions. Note that we produce unsigned char's
where Zenikos original produces ints, but it's otherwise
identical.

This requires us to make various knockon changes throughout the code
from int to unsigned char, none of which should make any difference
at all - I hope!

We reserve the right to change/amend this function in future to
operate on blocks of pixels (or pixels in different colorspaces)
(or anything else we think of) in future.