Support proper XPS mitering. (And stroke fixes).
XPS differs from PS/PDF/etc in the way it handles miters; rather than
simply converting a miter that's overly long to a bevel, it truncates
it at the miter limit. As such it needs to be handled correctly.

For clarity, expose new enumerated types for linejoins and linecaps,
and use these throughout code.

When we upgrade our freetype, we can move to using proper xps mitering
in that too.

Add new fz_matrix_max_expansion function to return a safer expansion
value that works in the case where we scale up in one direction and
down in another.

In the xps path drawing code, avoid generating unnecessary linetos.

Thanks to Zeniko for spotting these and providing implementations.