Fix possible SEGV/assert in downscaler.
On extreme downscales, the weights for each pixel can all round to
zero; this results in no weights being stored at all, and causes
either an exception or a SEGV. The simple fix is to ensure that the
first pixel that's actually in range (and there will always be one)
always has a weight of at least 1.

Thanks to malc for finding the test cases that lead to this bugfix.