Work on supporting links in xps documents.
Currently, this only works with local links.

When running the page, check for NavigateUri entries; if found,
and that page is not already marked as having resolved it's links,
add a new link entry to doc->current_page links. When the page
finishes running, mark the page as having resolved it's links.
This avoids the links being generated multiple times.

Update the mupdf viewer to use these links - but only AFTER the
page has been run.