Fix typo in hash table code.
When detecting a clash when inserting into a hash table (which
should only ever happen in multithreaded cases), we should give
a warning, and return the existing item from the table.

The current code doesn't do this due to a stupid typo. Fixed here.

CLUSTER_UNTESTED as we don't test multithreaded operation in the
cluster.