Fix incorrect handling of race condition.
When inserting an item into the store we check for an identically
keyed item being there already (for instance a pixmap created from
an image I at factor F may find that such a pixmap has already been
inserted). The correct thing to do is to return the old one so we
can use that in preference.

The code was attempting to do this, but was returning a pointer
to the fz_item rather than to the item->val. Fixed here.