Add -w, -h and -f options to mudraw.c
Exact width or height values can now be specified to mudraw for output
bitmaps. If both are specified the the largest scaling consistent
with the aspect ratio not exceeding the given width or height is used,
unless the -f flag is specified too, in which case the aspect ratio
is ignored and the output 'fits' exactly.