Bug 692990: Fix 'underjoin' of strokes.
See Bug 688655 for analysis of what we SHOULD be doing. The code changes
to do this are actually quite small.

Essentially, when we join we only draw the 'top' (or 'outer') join in a
join dependent way. The 'under' join was always joined as a bevel before
as this was easy. This produces bad effects when the lines have a
significant angle between them and a large linewidth.

The correct (i.e. matching Acrobat and others) way to work is to
join the bottom of the line via the centre point. The sole exception
to this is when drawing under beziers, as we don't want to make our
approximation-by-lines obvious.

All fixed in this patch.