Bug 692924: Allow mupdfinfo to accept files that don't end in .pdf
The logic in mupdfinfo was to interpret any command line argument
that ended in .pdf or .PDF as a filename, and to treat all others
as page range specifiers.

Here we change it so that page range specifiers are (roughly) detected,
and all other entries are taken to be filenames.

This doesn't address the wider issues of spotting files by 'magic' -
this is a larger task and will be addressed post 1.0.