Android tweak; more null checks.
When building MuPDFActivity as part of a framework, the application
lifecycle is slightly different. This can mean that if an attempt
is made to open a corrupt file (that fails to open) createUI can be
called with core == null. The fix is just to check that core is
non-null.

We add this check in various places to be doubly sure.