Fix crash in android app due to flood of page seeks.
Sebras found a way of crashing the MuPDF Android app by seeking
quickly back and forth with the page seek bar. After about 30
seconds of frantically tapping either end of it, we'd force close.

Examination shows this is due to the Async task queue getting full
and throwing an exception. A simple fix is to catch the exception
and perform the same tasks in the foreground.

Testing indicates that this causes the UI to stall, but it's far
preferable to a crash.