Bug 693022: Image/Pixmaps stuck in cache
Zeniko points out that images that don't decode on demand (i.e. ones
that are held as pixmaps all the time) can never be evicted from the
cache due to them holding a pointer to the pixmap, which holds a pointer
back to the image.

His fix is to only cache images that decode on demand.

The actual patch applied here is slightly tweaked from his version;
firstly the 'dontcache' logic is reversed to 'cache' to avoid overloading
my poor brain with another negation.

Secondly, one change to a condition is not taken on, as it is (I believe)
unnecessary.